+++ src/helpers.ts
@@ -1,5 +1,0 @@
-export async function getConfigDataFile(url: string, ttl: number) {
-  const fetchedDataFile = await fetch(url, { cf: { cacheTtl: ttl } });
-
-  return await fetchedDataFile.text();
-}